requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.1
defusedxml==0.7.1
docker==7.1.0
factory-boy==3.3.0
Faker==25.3.0
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpie==3.2.2
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
pluggy==1.5.0
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PySocks==1.7.1
pytest==8.2.0
pytest-cov==5.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.7.1
ruff==0.4.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
taskipy==1.12.2
testcontainers==4.5.1
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
zeep==4.2.1
